backend/upload-pdf/requirements.txt
boto3>=1.26.0
psycopg2-binary>=2.9.0
PyJWT>=2.8.0
cryptography>=41.0.0